Move fireworks.ts to proper folder
