+++ src/components/magicui/fireworks.ts
